Keylogger/requirements.txt
keyboard==0.13.5
Pillow==9.0.1
python-dotenv==0.19.0
requests==2.26.0
